+++ CHANGELOG.md
@@ -21,7 +21,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.12.0...master
+[Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.13.0...master
+
+## [3.13.0] - 2020-11-17
+[3.13.0]: https://github.com/atlassian/virtual-users/compare/release-3.12.0...release-3.13.0
 
 ### Added
 - Time VU tasks like throttling and diagnosis. Expose `VirtualUserResult.streamTasks`. Resolve [JPERF-483].